RELEASE BUILD!!!!!!!!!!!!!!!!!!!!!!

+++ src/commands/moderation/purge.ts
@@ -1,4 +1,4 @@
-import { ApplicationCommandOptionType, type GuildTextBasedChannel } from "discord.js";
+import { ApplicationCommandOptionType, DiscordAPIError, type GuildTextBasedChannel } from "discord.js";
 import type ICommand from "../../handler/interfaces/ICommand";
 
 export default {
@@ -18,11 +18,24 @@ export default {
                 ephemeral: true
             }
         }
-        await message?.delete();
-        const deletedMessages = await (channel as GuildTextBasedChannel).bulkDelete(amount, true);
-        return {
-            content: `Purged ${deletedMessages.size} messages`,
-            ephemeral: true
+
+        if (amount > 100) {
+            return {
+                content: "You can only purge 100 messages at a time",
+                ephemeral: true
+            }
+        }
+        try {
+            const deletedMessages = await (channel as GuildTextBasedChannel).bulkDelete(amount, true);
+            return {
+                content: `Purged ${deletedMessages.size} messages`,
+                ephemeral: true
+            }
+        } catch (e) {
+            return {
+                content: `Failed to purge messages (${(e as any).message})`,
+                ephemeral: true
+            }
         }
     },
